Split YoutubeApi classes into multiple files
