+++ lib/music_player/current_song_card/youtube_api/youtube_api.dart
@@ -1,133 +1,7 @@
 import 'dart:convert';
 
 import 'package:http/http.dart' as http;
-
-typedef YoutubeVideoId = String;
-typedef YoutubeChannelId = String;
-
-class YoutubeVideoThumbnails {
-  final YoutubeVideoThumbnail small;
-  final YoutubeVideoThumbnail medium;
-  final YoutubeVideoThumbnail high;
-  YoutubeVideoThumbnails({
-    required this.small,
-    required this.medium,
-    required this.high,
-  });
-
-  YoutubeVideoThumbnails copyWith({
-    YoutubeVideoThumbnail? small,
-    YoutubeVideoThumbnail? medium,
-    YoutubeVideoThumbnail? high,
-  }) {
-    return YoutubeVideoThumbnails(
-      small: small ?? this.small,
-      medium: medium ?? this.medium,
-      high: high ?? this.high,
-    );
-  }
-
-  Map<String, dynamic> toMap() {
-    return {
-      'default': small.toMap(),
-      'medium': medium.toMap(),
-      'high': high.toMap(),
-    };
-  }
-
-  factory YoutubeVideoThumbnails.fromMap(Map<String, dynamic> map) {
-    return YoutubeVideoThumbnails(
-      small: YoutubeVideoThumbnail.fromMap(map['default']),
-      medium: YoutubeVideoThumbnail.fromMap(map['medium']),
-      high: YoutubeVideoThumbnail.fromMap(map['high']),
-    );
-  }
-
-  String toJson() => json.encode(toMap());
-
-  factory YoutubeVideoThumbnails.fromJson(String source) =>
-      YoutubeVideoThumbnails.fromMap(json.decode(source));
-
-  @override
-  String toString() =>
-      'Thumbnails(default: $small, medium: $medium, high: $high)';
-
-  @override
-  bool operator ==(Object other) {
-    if (identical(this, other)) return true;
-
-    return other is YoutubeVideoThumbnails &&
-        other.small == small &&
-        other.medium == medium &&
-        other.high == high;
-  }
-
-  @override
-  int get hashCode => small.hashCode ^ medium.hashCode ^ high.hashCode;
-}
-
-class YoutubeVideoThumbnail {
-  const YoutubeVideoThumbnail({
-    required this.url,
-    required this.width,
-    required this.height,
-  });
-
-  final String url;
-
-  final int width;
-  final int height;
-
-  Map<String, dynamic> toMap() {
-    return {
-      'url': url,
-      'width': width,
-      'height': height,
-    };
-  }
-
-  factory YoutubeVideoThumbnail.fromMap(Map<String, dynamic> map) {
-    return YoutubeVideoThumbnail(
-      url: map['url'],
-      width: map['width'],
-      height: map['height'],
-    );
-  }
-
-  String toJson() => json.encode(toMap());
-
-  factory YoutubeVideoThumbnail.fromJson(String source) =>
-      YoutubeVideoThumbnail.fromMap(json.decode(source));
-}
-
-class YoutubeVideo {
-  final YoutubeVideoId id;
-  final String url;
-
-  final YoutubeChannelId channelId;
-  final String channelUrl;
-  final String channelTitle;
-
-  final String title;
-  final String description;
-  final DateTime publishedAt;
-
-  final YoutubeVideoThumbnails thumbnails;
-
-  YoutubeVideo.fromJson(dynamic data)
-      : thumbnails =
-            YoutubeVideoThumbnails.fromMap(data['snippet']['thumbnails']),
-        id = data['id'][data['id'].keys.elementAt(1)],
-        url =
-            "https://www.youtube.com/watch?v=${data['id'][data['id'].keys.elementAt(1)]}",
-        publishedAt = DateTime.parse(data['snippet']['publishedAt']),
-        channelId = data['snippet']['channelId'],
-        channelUrl =
-            "https://www.youtube.com/channel/${data['snippet']['channelId']}",
-        title = data['snippet']['title'],
-        description = data['snippet']['description'],
-        channelTitle = data['snippet']['channelTitle'];
-}
+import 'package:musbx/music_player/current_song_card/youtube_api/video.dart';
 
 class YoutubeApi {
   static String baseURL = 'www.googleapis.com';
