+++ lib/music_player/current_song_card/youtube_button.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:html_unescape/html_unescape.dart';
 import 'package:musbx/custom_icons.dart';
 import 'package:musbx/music_player/api_key.dart';
+import 'package:musbx/music_player/current_song_card/youtube_api/video.dart';
 import 'package:musbx/music_player/current_song_card/youtube_api/youtube_api.dart';
 import 'package:musbx/music_player/music_player.dart';
 import 'package:musbx/widgets.dart';
