+++ lib/music_player/music_player.dart
@@ -7,8 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:html_unescape/html_unescape.dart';
 import 'package:just_audio/just_audio.dart';
 import 'package:musbx/music_player/audio_handler.dart';
-import 'package:musbx/music_player/current_song_card/youtube_api/youtube_api.dart';
-import 'package:youtube_api/youtube_api.dart';
+import 'package:musbx/music_player/current_song_card/youtube_api/video.dart';
 import 'package:youtube_explode_dart/youtube_explode_dart.dart';
 
 /// The state of [MusicPlayer].
